requirements.txt
cairosvg==2.5.2
django-allauth~=0.44
django-comments-xtd==2.9.*
django-compressor==2.4.*
django-debug-toolbar==3.2.1
django-extensions==3.1.*
django-sass-processor==1.0.*
django-translation-manager==1.1.1
django==3.1.*
elasticsearch~=7.16
libsass==0.21.*
lxml==4.8.*
psycopg2==2.8.*
redis~=3.0
tqdm==4.62.*
wagtail-localize==1.1.1
wagtail-markdown==0.7.0
Markdown==3.3.7
wagtail==2.11.*
wagtailmedia==0.7.0
wagtailmenus~=3.0
wagtailsvg==0.0.14
whitenoise==5.2.*
django-health-check==3.16.5
djangorestframework-simplejwt==5.0.0
django-redis==5.2.0
django-google-analytics-app==5.0.5
django-filter==2.4.0
git+https://github.com/IDEMSInternational/wagtail-transfer.git@enh/show-error-for-unimported-referenced-pages#egg=wagtail-transfer
